Decouple assertion logic and add log settings
Add the possibility to configure the logs output with
`LogSettings`. This commit also decouple assertion logic from
the DSL as much as possible.
Each assertion can now produce its own output in a realtime
fashion. This will allow future implementations of log drivers
like elasticsearch.

This new logic has been partially implemented, and at the moment
only status code can produce different kind of logs. Subsequent
commits will extend the behaviour to other http parts.

This commit also changes the way we calculate the response time.
We now use Instant::now() for a better precision. However we now
use u64 instead of u128 for response time milliseconds. This is
due to some limitation in serde.

+++ src/dsl/http/time.rs
@@ -20,8 +20,8 @@ pub trait TimeDsl<T> {
     }
 }
 
-impl TimeDsl<u128> for u128 {
-    fn is_less_than(&self, actual: u128) -> Assertion {
+impl TimeDsl<u64> for u64 {
+    fn is_less_than(&self, actual: u64) -> Assertion {
         let result = actual < *self;
 
         let ty = AssertBool {
